requirements.txt
Django
git+https://github.com/TheEnbyperor/django-keycloak-auth@8f8b3f8a6e554acb7aff21ab12ec54c6b59b5197
sentry-sdk
psycopg2-binary
gunicorn
django-crispy-forms
django-xff
retry
requests
django-cockroachdb==3.1.*
git+https://github.com/AS207960/python-utils@e345a3168b7040fbf5481e127bfa7cb78e2a8dcc
python-jose[cryptography]